Another refactoring
The goal was, to have the same internal state of each signal. A distant
signal is now just a different representation of the same state.

An alternative would have been to link the distant signals to the main
signal. I tried that idea, but it created quite a bit of uncertainty
(many Maybe's) and was a really big change. Also it wouldn't have been
possible to just have a distant signal without an (invisible) main
signal.